requirements.txt
flask
requests
gunicorn
azure-identity
azure-keyvault-secrets
